Added a test for solve3.
solve3 is intended to reduce the number of greenhouses even further
by first of all enforcing the max constraint by joining adjacent
greenhouses with the least number of empty spaces across them. Once
it comes down to the constraint it continues to join greenhouses
if it lowers the overall cost.