Fixed the cost function. Also extracted ids function from solve3.
Turns out I needed the list of greenhouse identities in other parts
of the program as well so I extracted it. Cost was fairly straight
to implement given the helper functions available. I may have to wrap
some of the in memoizers to get up to speed but that's a later worry.