Finisihed the solve method.
This leaves us with a program that actually can produce a solution.
It will most probably not produce the optimal solution just now
because it only makes sure to reduce the number of greenhouses
until the max constraint is met. Further reductions can be tried.

I'll start working on that solve4 function now. Refactoring comes
a little later.