Added one more puzzle.
This is the same as the previous one, only rotated 90 degrees.
Unfortunately the solver chain comes up with different solutions
to these two. Not that suprising given how the solve2 method
works but it clearly points out the problem with not having a
search algorithm.

Also, the 'horizontal bias' in solve2 means that we end up with
suboptimal solutions. That will have to go into the refactoring
stuff later on.