Minor cleanup in solve2.
I'm going to scrap the constraint that the row below needs to
look exactly like the one we just found and have it instead
merge as greedily as possible. Not only will that remove some
lines of code but I think it will end up with a better set of
greenhouses for solve3 as well.