Added another test for solve2.
I'm going to redo how solve2 works. I do not want it to join
greenhouses horizontally across 'edges' or 'zones'. Meaning
that if you look at the cells above and below they should not
change in order for the cell to be joined. This means we'll
have a lot more greenhouses after solve2 but we will be in a
much better position to do the subsequent merges.