Added a test to outer_bounds.
Still trying to locate the problem with solve4 generating
faulty solutions. One way to go about it is to have
outer_bounds behave differently, returning None instead of
a calculated size if the outer_bounds contain any cell
not found in the list of greenhouse ids to look for. This
way, solve4 can reject faulty joins.