Added more variations of the puzzle.
Since the simple reduction used is working from left to right, up to
down. It will end up with different solutions if you rotate and/or
flip the field. A simple way of improving the results is to produce
all 8 variations of a given puzzle and use the one with the lowest
cost.