Added a test and function def for flipping fields horizontally.
Running simple_reduction on p1.text which contains all 8 variations
of the same puzzle reveals that we can get a better solution if we
explore the puzzle from different angles.

I intend to have a reduciton method that simple feeds all variations
to simple_reduction and returns the best one.