Fixed the problem highlighted by p4.text.
Now, we've come to the point where it's possible to say
that we cannot find a solution. we do that by returning
None instead of a field. In order for that to work in the
solve chain I updated the cost function to return
sys.maxint when it encounters None.