Modified the search algorithm to use simple_reduction as a heuristic.
Running the previous s2 code on some of the other puzzles was
showing really poor results. I aborted a test on p1.text after
ten minutes, the version I'm commiting now returns after one
minute and 15 seconds. Compared with s1 which returns after four
seconds it is of course not very good but I hope to be able to
improve this.